nodelet: fix segfault in devicePoll().
device poll can be called prior to pub_ instantiation, since the connection
callback is created prior to the publisher ctor in onInit().

null-check the publisher in the connection callback to avoid triggering
segfaults when the race condition is triggered.

note, this doesn't fix the race condition, since that'd require messing with
the connection callback setup, which is required for creating the publisher,
but we shouldn't need to resolve that in any other way.

this normally happened when starting a bunch of other ros nodes along with the
oryx node.